# Prefill Packing: refactor code into a python module