# Description

**Goal**

Extract prefill packing logic
- from MLPerf `offline_inference.py` and `maxengine.py`
- to `prefill_packing.py`

**Performance Stats**

Before
- MLPerf + OfflineInference 3388 tokens/s
- MLPerf + OfflineInference (with prefill packing) 3869 tokens/s

After
- MLPerf + OfflineInference 3403 tokens/s
- MLPerf + OfflineInference (with prefill packing) 3880 tokens/s

**Details**

offline_inference.py
- OfflineInference
    - del `prefill API & prefill-packing API aot`: line 100-177, line 181-221
    - del `prefill API & prefill-packing API call`: line 234-314
        - OfflineInference calls a unified prefill API provided by PrefillHelper.
    - del `prefill pack & unpack`: line 234-314, line 394-459
    - del `prefill collect`: line 394-459
- PrefillHelper
    - manages BatchedPrefillProcessor and PrefillProcessor.
    - provides a unified interface for OfflineInference.

prefill_packing.py
- PrefillBucket
    - add `prefill collect`.
- BatchedPrefillProcessor
    - add `prefill-packing API`.
    - add `prefill-packing API aot`.
    - add `prefill-packing API call`.
    - add `prefill pack & unpack`.
- PrefillProcessor
    - add `prefill API aot`.
    - add `prefill API call`.

# Tests

MLPerf

```
$ cd maxtext/MaxText/inference_mlperf
$ DATA_DISK_DIR=$HOME/tmp/ \
TOKENIZER_PATH=$HOME/github/maxtext/assets/tokenizer.llama2 \
BASEDIR=$HOME/github/inference \
PREFILL_LENS_AND_PER_DEVICE_BATCH_SIZES=1024,64 \
MAXENGINE_ARGS="model_name=llama2-70b \
    tokenizer_path=${TOKENIZER_PATH} \
    load_parameters_path=$HOME/model/direct_generate_param_only_checkpoint_llama2_70b_chat/checkpoints/0/items \
    quantize_kvcache=False \
    attention=dot_product \
    use_ragged_attention=false \
    scan_layers=false \
    hardware=gpu \
    async_checkpointing=False \
    ici_fsdp_parallelism=1 \
    ici_autoregressive_parallelism=-1 \
    ici_tensor_parallelism=8 \
    weight_dtype=bfloat16" \
./llama_offline_run.sh
```

MLPerf (enable prefill packing)

```
$ cd maxtext/MaxText/inference_mlperf
$ DATA_DISK_DIR=$HOME/tmp/ \
TOKENIZER_PATH=$HOME/github/maxtext/assets/tokenizer.llama2 \
BASEDIR=$HOME/github/inference \
PREFILL_LENS_AND_PER_DEVICE_BATCH_SIZES=1024,64 \
MAXENGINE_ARGS="model_name=llama2-70b \
    tokenizer_path=${TOKENIZER_PATH} \
    load_parameters_path=$HOME/model/direct_generate_param_only_checkpoint_llama2_70b_chat/checkpoints/0/items \
    quantize_kvcache=False \
    attention=dot_product \
    use_ragged_attention=false \
    scan_layers=false \
    hardware=gpu \
    async_checkpointing=False \
    ici_fsdp_parallelism=1 \
    ici_autoregressive_parallelism=-1 \
    ici_tensor_parallelism=8 \
    weight_dtype=bfloat16" \
./llama_offline_run.sh -c
```

# Checklist

Before submitting this PR, please make sure (put X in square brackets):
- [x] I have performed a self-review of my code.
- [x] I have necessary comments in my code, particularly in hard-to-understand areas.
- [x] I have run end-to-end tests tests and provided workload links above if applicable.
- [x] I have made or will make corresponding changes to the doc if needed.