@wang2yn84 2025-04-14
2 things I'd like to point out: 1. I can only run the script with "export PYTHONPATH="$(realpath $(dirname $0)/../..):$PYTHONPATH"" added to the llama_offline_run.sh, otherwise it complains about not being able to fine MaxText module. 2. The best bf16 config is batch size 80, maybe should add XLA_PYTHON_CLIENT_MEM_FRACTION=0.94 to the command description of the PR.