@vipannalla 2025-04-14
Thanks for refactoring, for "MLPerf + OfflineInference (with prefill packing) 3880 tokens/s" -- is this on Trillium or GPU? That number is pretty low, the current numbers are around ~6k.  Can you share you command?
@mitalisi , @sixiang-google PTAL.